рефакторит Reducer. WIP
