+++ project/src/store/user-process/user-process.ts
@@ -1,0 +1,21 @@
+import {Actions, ActionType} from '../../types/action';
+import {AuthorizationStatus} from '../../const';
+import {UserProcess} from '../../types/state';
+
+
+const initialState: UserProcess = {
+  authorizationStatus: AuthorizationStatus.Unknown,
+};
+
+const userProcess = (state = initialState, action: Actions): UserProcess => {
+  switch (action.type) {
+    case ActionType.RequireAuthorization:
+      return {...state, authorizationStatus: action.payload};
+    case ActionType.RequireLogout:
+      return {...state, authorizationStatus: AuthorizationStatus.NoAuth};
+    default:
+      return state;
+  }
+};
+
+export {userProcess};
