+++ project/src/types/state.ts
@@ -3,12 +3,19 @@ import {Genres, AuthorizationStatus} from '../const';
 import {UserType} from './user';
 import {CommentsType} from './comment';
 
-export type State = {
-  genre: Genres,
+
+export type AppData = {
   movies: MoviesType,
+  user: UserType,
+}
+
+export type MovieProcess = {
+  genre: Genres,
   movie: MovieType | null,
   comments: CommentsType,
   similarMovies: MoviesType,
+}
+
+export type UserProcess = {
   authorizationStatus: AuthorizationStatus,
-  user: UserType,
-};
+}
